Create server main loop
