+++ git-keeper-server/gkeepserver/email_sender_thread.py
@@ -46,7 +46,7 @@ from threading import Thread
 from queue import Queue, Empty
 from time import time, sleep
 
-from gkeepserver.email import Email, EmailException
+from gkeepserver.server_email import Email, EmailException
 
 
 class EmailSenderThread(Thread):
@@ -122,7 +122,7 @@ class EmailSenderThread(Thread):
         while not self._shutdown_flag:
             try:
                 while True:
-                    email = self._email_queue.get(block=True, timeout=0.5)
+                    email = self._email_queue.get(block=True, timeout=0.1)
 
                     if isinstance(email, Email):
                         self._send_email_with_rate_limiting(email)
@@ -154,3 +154,7 @@ class EmailSenderThread(Thread):
         except EmailException as e:
             # FIXME - log this instead
             print('EMAILER: Failed to send email to', email.to_address)
+
+
+# module-level instance for global access
+email_sender = EmailSenderThread()
